requirements.txt
pandas
pytube
SpeechRecognition
pydub
PyYAML==5.4.1
python-box